Feature(Integration): for Mailchimp, moving util functions to different file
